DEBUG: attempt to reduce memory heap usage: add more context types for usage between components, broke chatroom components into multiple subcomponents that prevent rerender
