+++ src/client/contexts/ChatroomContext.tsx
@@ -2,5 +2,7 @@ import React from "react";
 
 const ChatroomContext = React.createContext();
 const ToggleDJContext = React.createContext();
+const SocketContext = React.createContext();
+const DataContext = React.createContext(null);
 
-export { ChatroomContext, ToggleDJContext };
+export { ChatroomContext, ToggleDJContext, SocketContext, DataContext };
