+++ src/client/views/Chatroom.tsx
@@ -11,7 +11,6 @@ import React, {
 } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import io from 'socket.io-client';
-import { Application, extend, useAssets } from '@pixi/react';
 import dayjs from 'dayjs';
 import { Label } from '../../components/ui/label';
 import { AnimatedList } from '../../components/ui/animated-list';
@@ -27,44 +26,14 @@ const MsgBox = lazy(() => import('../components/chatroom/MsgBox'));
 const DJam = lazy(() => import('../components/chatroom/DJam'));
 const Menu = lazy(() => import('../components/chatroom/Menu'));
 import SOCKET_URL from '../../../config';
-import TILES from '../assets/chatroom/tiles/index';
-import IDLE from '../assets/chatroom/idle/index';
-import WALK from '../assets/chatroom/walk/index';
-import SNAP from '../assets/chatroom/snap/index';
-import WAVE from '../assets/chatroom/wave/index';
-import EMOJIS from '../assets/chatroom/emojis';
-import speechbubble from '../assets/images/speechbubble.png'
 import loading from '../assets/chatroom/loading.gif';
-import ENERGYWAVE from '../assets/chatroom/energy/index';
-import mapPack from '../assets/chatroom/mapPack';
+
 import { FaShip } from 'react-icons/fa';
-
-import {
-  Container,
-  Graphics,
-  Sprite,
-  Texture,
-  Assets,
-  NineSliceSprite, // failing
-  Text,
-  TextStyle,
-  AnimatedSprite,
-} from 'pixi.js';
+const MainChat = lazy(() => import('../components/chatroom/MainChat'));
 import axios from 'axios';
-import nightClubTileSet from '../assets/chatroom/tileSet';
-import { ChatroomContext, ToggleDJContext } from '../contexts/ChatroomContext';
 
-extend({
-  Container,
-  Graphics,
-  Sprite,
-  Texture,
-  NineSliceSprite,
-  Text,
-  TextStyle,
-  AnimatedSprite,
-  Texture, // not worth it w/ useAssets...?
-});
+import { ChatroomContext, ToggleDJContext, SocketContext } from '../contexts/ChatroomContext';
+
 
 let socket = io(SOCKET_URL);
 
@@ -73,236 +42,32 @@ function Chatroom() {
   const { user } = useContext(UserContext);
   const location = useLocation();
   const start_time = location.state;
-
-  // LOAD ASSETS
-  const [gameLoaded, setGameLoaded] = useState(false);
-  const [texturesToLoad, settexturesToLoad] = useState([
-    { alias: 'bunny', src: 'https://pixijs.com/assets/bunny.png' },
-     {
-       alias: 'speech',
-       src: speechbubble,
-     },
-    { alias: 'joint', src: EMOJIS['7'] }, // joint,
-    { alias: '1', src: TILES['1'] },
-    { alias: '2', src: TILES['2'] },
-    { alias: '3', src: TILES['3'] },
-    { alias: '4', src: TILES['4'] },
-    { alias: '5', src: TILES['5'] },
-    { alias: '6', src: TILES['6'] },
-    { alias: '7', src: TILES['7'] },
-    { alias: '8', src: TILES['8'] },
-    { alias: '9', src: TILES['9'] },
-    { alias: '10', src: TILES['10'] },
-    { alias: '11', src: TILES['11'] },
-    { alias: '12', src: TILES['12'] },
-    { alias: '13', src: TILES['13'] },
-    { alias: '14', src: TILES['14'] },
-    { alias: '15', src: TILES['15'] },
-    { alias: '16', src: TILES['16'] },
-    { alias: '17', src: TILES['17'] },
-    { alias: '18', src: TILES['18'] },
-    { alias: '19', src: TILES['19'] },
-    { alias: '20', src: TILES['20'] },
-    { alias: '21', src: TILES['21'] },
-    { alias: '22', src: TILES['22'] },
-    { alias: '23', src: TILES['23'] },
-    { alias: '24', src: TILES['24'] },
-    { alias: '25', src: TILES['25'] },
-    { alias: '26', src: TILES['26'] },
-    { alias: '27', src: TILES['27'] },
-    { alias: '28', src: TILES['28'] },
-    { alias: '29', src: TILES['29'] },
-    { alias: '30', src: TILES['30'] },
-    { alias: '31', src: TILES['31'] },
-    { alias: '32', src: TILES['32'] },
-    { alias: '33', src: TILES['33'] },
-    { alias: '34', src: TILES['34'] },
-    { alias: '35', src: TILES['35'] },
-    { alias: '36', src: TILES['36'] },
-    { alias: '37', src: TILES['37'] },
-    { alias: '38', src: TILES['38'] },
-    { alias: '39', src: TILES['39'] },
-    { alias: '40', src: TILES['40'] },
-    { alias: '41', src: TILES['41'] },
-    { alias: '42', src: TILES['42'] },
-    { alias: '43', src: TILES['43'] },
-    { alias: '44', src: TILES['44'] },
-    { alias: '45', src: TILES['45'] },
-    { alias: '46', src: TILES['46'] },
-    { alias: '47', src: TILES['47'] },
-    { alias: '48', src: TILES['48'] },
-    { alias: '49', src: TILES['49'] },
-    { alias: '50', src: TILES['50'] },
-    { alias: '51', src: TILES['51'] },
-    { alias: '52', src: TILES['52'] },
-    { alias: '53', src: TILES['53'] },
-    { alias: '54', src: TILES['54'] },
-    { alias: '55', src: TILES['55'] },
-    { alias: '56', src: TILES['56'] },
-    { alias: '57', src: TILES['57'] },
-    { alias: '58', src: TILES['58'] },
-    { alias: '59', src: TILES['59'] },
-    { alias: '60', src: TILES['60'] },
-    { alias: '61', src: TILES['61'] },
-    { alias: '62', src: TILES['62'] },
-    { alias: '63', src: TILES['63'] },
-    { alias: '64', src: TILES['64'] },
-    { alias: '65', src: TILES['65'] },
-    { alias: '66', src: TILES['66'] },
-    { alias: '67', src: TILES['67'] },
-    { alias: '68', src: TILES['68'] },
-    { alias: '69', src: TILES['69'] },
-    { alias: '70', src: TILES['70'] },
-    { alias: '71', src: TILES['71'] },
-    { alias: '72', src: TILES['72'] },
-    { alias: '73', src: TILES['73'] },
-    { alias: '74', src: TILES['74'] },
-    { alias: '75', src: TILES['75'] },
-    { alias: '76', src: TILES['76'] },
-    { alias: '77', src: TILES['77'] },
-    { alias: '78', src: TILES['78'] },
-    { alias: '79', src: TILES['79'] },
-    { alias: '80', src: TILES['80'] },
-    { alias: '81', src: TILES['81'] },
-    { alias: '82', src: TILES['82'] },
-    { alias: '83', src: TILES['83'] },
-    { alias: '84', src: TILES['84'] },
-    { alias: '85', src: TILES['85'] },
-    { alias: '86', src: TILES['86'] },
-    { alias: '87', src: TILES['87'] },
-    { alias: '88', src: TILES['88'] },
-    { alias: '89', src: TILES['89'] },
-    { alias: '90', src: TILES['90'] },
-    { alias: '91', src: TILES['91'] },
-    { alias: '92', src: TILES['92'] },
-    { alias: '93', src: TILES['93'] },
-    { alias: '94', src: TILES['94'] },
-    { alias: '95', src: TILES['95'] },
-    { alias: '96', src: TILES['96'] },
-    { alias: '97', src: TILES['97'] },
-    { alias: '98', src: TILES['98'] },
-    { alias: '99', src: TILES['99'] },
-    { alias: '100', src: TILES['100'] },
-    { alias: '101', src: TILES['101'] },
-    { alias: '102', src: TILES['102'] },
-    { alias: '103', src: TILES['103'] },
-    { alias: '104', src: TILES['104'] },
-    { alias: '105', src: IDLE['105'] },
-    { alias: '106', src: IDLE['106'] },
-    { alias: '107', src: IDLE['107'] },
-    { alias: '108', src: IDLE['108'] },
-    { alias: '109', src: IDLE['109'] },
-    { alias: '110', src: WALK['110'] },
-    { alias: '111', src: WALK['111'] },
-    { alias: '112', src: WALK['112'] },
-    { alias: '113', src: WALK['113'] },
-    { alias: '114', src: WALK['114'] },
-    { alias: '115', src: WALK['115'] },
-    { alias: '116', src: SNAP['1'] },
-    { alias: '117', src: SNAP['2'] },
-    { alias: '118', src: SNAP['3'] },
-    { alias: '119', src: SNAP['4'] },
-    { alias: '120', src: SNAP['5'] },
-    { alias: '121', src: WAVE['1'] },
-    { alias: '122', src: WAVE['2'] },
-    { alias: '123', src: WAVE['3'] },
-    { alias: '124', src: WAVE['4'] },
-    { alias: '125', src: WAVE['5'] },
-    { alias: '126', src: ENERGYWAVE['1'] },
-    { alias: '127', src: ENERGYWAVE['2'] },
-    { alias: '128', src: ENERGYWAVE['3'] },
-    { alias: '129', src: ENERGYWAVE['4'] },
-    { alias: '130', src: ENERGYWAVE['5'] },
-    { alias: '131', src: ENERGYWAVE['6'] },
-    { alias: '132', src: ENERGYWAVE['7'] },
-    { alias: '133', src: ENERGYWAVE['8'] },
-    { alias: '134', src: EMOJIS['1'] },
-    { alias: '135', src: EMOJIS['2'] },
-    { alias: '136', src: EMOJIS['3'] },
-    { alias: '137', src: EMOJIS['4'] },
-    { alias: 'sad', src: EMOJIS['5'] }, //sad
-    { alias: 'shades', src: EMOJIS['6'] }, ///shades
-    { alias: 'beer', src: EMOJIS['8'] }, // beer
-  ]);
-  const { assets, isSuccess } = useAssets(texturesToLoad);
-
+  
+  const [avatarTextures, setAvatarTextures] = useState<{ alias: any; src: any; }[]>([])
   const [lobby, setLobby] = useState([user]);
   const [playerY, setPlayerY] = useState(0);
   const [playerX, setPlayerX] = useState(0);
   const [playerPosition, setPlayerPosition] = useState([playerY, playerX]);
-
   // LOGIC
-  const appRef = useRef(null);
-  const [gameRatio, setGameRatio] = useState(
-    window.innerWidth / window.innerHeight
-  );
-  const [scaleFactor, setScaleFactor] = useState(gameRatio > 1.5 ? 0.8 : 1);
-  const [allPlayers, setAllPlayers] = useState([]);
+  
+  
+
   const [eventId, setEventId] = useState(document.location.pathname.slice(10));
   const [message, setMessage] = useState('');
   const [isTyping, setIsTyping] = useState(false);
   const [allMessages, setAllMessages] = useState([]);
-  const spriteRef = useRef(null);
-  const spriteRef2 = useRef(null);
+  
   const [isPlayingFlamiliar, setIsPlayingFlamiliar] = useState(false);
   const [isPlayingDJ, setIsPlayingDJ] = useState(false);
   const [isPlayingGames, setIsPlayingGames] = useState(false);
   const displayMessage = (msg: any) => {
     setAllMessages((prevMessages) => [...prevMessages, msg]);
   };
-  const [textures, setTextures] = useState([]);
-  const [walkTextures, setWalkTextures] = useState([]);
-  const [snapTextures, setSnapTextures] = useState([]);
-  const [waveTextures, setWaveTextures] = useState([]);
-  const [heartTextures, setHeartTextures] = useState([]);
-  const [energyWaveTextures, setEnergyWaveTextures] = useState([]);
+ 
   const [onKeyboard, setOnKeyboard] = useState<boolean>(false);
-  const [isReady, setIsReady] = useState(false);
 
-  const styleMessage = new TextStyle({
-    align: 'center',
-    fontFamily: 'sans-serif',
-    fontSize: 10,
-    fontWeight: 'bold',
-    fill: '#000000',
-    stroke: '#eef1f5',
-    letterSpacing: 2,
-    wordWrap: true,
-    wordWrapWidth: 80,
-  });
 
-  const styleUserName = new TextStyle({
-    align: 'center',
-    fontFamily: 'sans-serif',
-    fontSize: 15,
-    fontWeight: 'bold',
-    fill: '#000000',
-    stroke: '#eef1f5',
-    letterSpacing: 5,
-    wordWrap: true,
-    wordWrapWidth: 250,
-  });
-
-  useEffect(() => {
-    socket.on('newPlayerList', ({ PLAYER_LIST }) => {
-      for (let player in PLAYER_LIST) {
-        if (!lobby.includes(PLAYER_LIST[player].username)) {
-          setLobby((prevItems) => [...prevItems, PLAYER_LIST[player].username]);
-          settexturesToLoad((prevItems) => [
-            ...prevItems,
-            {
-              alias: PLAYER_LIST[player].username,
-              src: PLAYER_LIST[player].avatar,
-            },
-          ]);
-        }
-      }
-    });
-    return () => {
-      socket.off('newPlayerList');
-    };
-  }, []);
-
+  
   // Flamiliar
   const toggleGames = () => {
     isPlayingGames ? setIsPlayingGames(false) : setIsPlayingGames(true);
@@ -381,68 +146,32 @@ function Chatroom() {
       socket.emit('keyPress', { inputId: 'Sad', state: false });
     }
   };
-  // SOCKET ACTIVITY & MAP LOAD
-  useEffect(() => {
-    axios.get(`api/chatroom/${eventId}`).catch((err) => console.error(err));
-    socket.emit('joinChat', { user, eventId });
-    socket.on('message', (msg) => {
-      displayMessage(msg);
-    });
-    socket.on('newPositions', (data) => {
-      let allPlayerInfo = [];
-      for (let i = 0; i < data.length; i++) {
-        if (data[i].room === eventId) {
-          allPlayerInfo.push({
-            id: data[i].id,
-            avatar: data[i].avatar,
-            x: data[i].x,
-            y: data[i].y,
-            username: data[i].username,
-            sentMessage: data[i].sentMessage,
-            currentMessage: data[i].currentMessage,
-            room: data[i].room,
-            isWalking: data[i].isWalking,
-            isSnapping: data[i].isSnapping,
-            isWaving: data[i].isWaving,
-            isEnergyWaving: data[i].isEnergyWaving,
-            isHearting: data[i].isHearting,
-            equipShades: data[i].equipShades,
-            equip420: data[i].equip420,
-            equipBeer: data[i].equipBeer,
-            isSad: data[i].isSad,
-          });
 
-          // if the client is at the keyboard:
-          if (data[i].username === user.username) {
-            if (
-              data[i].x < 466 &&
-              data[i].x > 412 &&
-              data[i].y > 112 &&
-              data[i].y < 150 &&
-              !onKeyboard
-            ) {
-              setOnKeyboard(true);
-            }
 
-            if (
-              data[i].x > 466 ||
-              data[i].x < 412 ||
-              ((data[i].y < 112 || data[i].y > 150) && onKeyboard)
-            ) {
-              if(onKeyboard){
-              setOnKeyboard(false);
+        // SOCKET ACTIVITY & MAP LOAD
+        useEffect(() => {
+          console.log('new player list retriggers')
+          socket.on('newPlayerList', ({ PLAYER_LIST }) => {
+            console.log(PLAYER_LIST)
+            console.log(lobby)
+            for (let player in PLAYER_LIST) {
+              if (!lobby.includes(PLAYER_LIST[player].username)) {
+                setLobby((prevItems) => [...prevItems, PLAYER_LIST[player].username]);
+                setAvatarTextures((prevItems) => [
+                  ...prevItems,
+                  {
+                    alias: PLAYER_LIST[player].username,
+                    src: PLAYER_LIST[player].avatar,
+                  },
+                ]);
               }
             }
+          });
+   
+          return () => {
+            socket.off('newPlayerList')
           }
-        }
-      }
-      setAllPlayers(allPlayerInfo);
-    });
-    return () => {
-      socket.off('newPositions');
-      socket.off('message');
-    };
-  }, []);
+        }, [avatarTextures]);
   // EVENT LISTENERS FOR TYPING
   useEffect(() => {
     if (isTyping === false) {
@@ -463,79 +192,8 @@ function Chatroom() {
     await setIsTyping(true);
   };
   //////////////////////////////////////////////////
-
-  useEffect(() => {
-    if (spriteRef.current && isReady) {
-      spriteRef.current.play(); // Explicitly start animation
-    }
-  }, []);
   useEffect(() => {}, [setOnKeyboard]);
-  useEffect(() => {
-    return () => {
-      // Cleanup sprite texture when unmounting
-      if (spriteRef2.current) {
-        spriteRef2.current.texture?.destroy(true);
-      }
-    };
-  }, []);
-
-  useEffect(() => {
-    if (isSuccess) {
-      const loadedTextures = [
-        assets['107'],
-        assets['108'],
-        assets['109'],
-        assets['110'],
-        assets['111'],
-      ];
-      const loadedWalkTextures = [
-        assets['112'],
-        assets['113'],
-        assets['114'],
-        assets['115'],
-        assets['116'],
-        assets['117'],
-      ];
-      const loadedSnapTextures = [
-        assets['118'],
-        assets['119'],
-        assets['120'],
-        assets['121'],
-        assets['122'],
-      ];
-      const loadedWaveTextures = [
-        assets['123'],
-        assets['124'],
-        assets['125'],
-        assets['126'],
-        assets['127'],
-      ];
-      const loadedEnergyWaveTextures = [
-        assets['128'],
-        assets['129'],
-        assets['130'],
-        assets['131'],
-        assets['132'],
-        assets['133'],
-        assets['134'],
-        assets['135'],
-      ];
-      const loadedHeartTextures = [
-        assets['136'],
-        assets['137'],
-        assets['138'],
-        assets['139'],
-      ];
-      setHeartTextures(loadedHeartTextures);
-      setEnergyWaveTextures(loadedEnergyWaveTextures);
-      setTextures(loadedTextures);
-      setWalkTextures(loadedWalkTextures);
-      setSnapTextures(loadedSnapTextures);
-      setWaveTextures(loadedWaveTextures);
-      setIsReady(true); // Once textures are ready, set the state to true
-    }
-  }, [isSuccess, assets]); // Re-run when assets load
-
+ 
   ///////////////////////////////////////////
   const notTyping = async () => {
     await setIsTyping(false);
@@ -550,17 +208,22 @@ function Chatroom() {
     setMessage('');
   };
 
-  // WINDOW SIZING
-  const handleResize = () => {
-    setGameRatio(window.innerWidth / window.innerHeight);
-    setScaleFactor(gameRatio > 1.3 ? 0.75 : 1);
-  };
+useEffect(() => {
+  console.log('message retriggers')
+  socket.emit('joinChat', { user, eventId });
+  axios.get(`api/chatroom/${eventId}`).catch((err) => console.error(err));
+  socket.on('message', (msg) => {
+    displayMessage(msg);
+  });
 
-  useEffect(() => {
-    window.addEventListener('resize', handleResize);
+  return () => {
+    socket.off('message');
+  }
+}, [])
 
-    return () => window.removeEventListener('resize', handleResize);
-  }, []);
+const chatSetOnKeyboard = (val) => {
+    setOnKeyboard(val);
+}
 
   const handlePointerDown = (e) => {
     keyPress({ key: e.target.name }); // Adjust the timeout as needed
@@ -571,9 +234,7 @@ function Chatroom() {
   };
   return (
     <div className="min-h-screen bg-gradient-to-br from-black via-gray-900 to-pink-900 relative overflow-hidden">
-      <div>
-        <img src={TILES['1']} alt="" />
-      </div>
+     
       <div
         className={`${isPlayingGames ? 'invisible': ''} flex justify-center mt-6`}
       >
@@ -628,7 +289,7 @@ function Chatroom() {
                               alt="Loading..."
                             ></img>
                             </div>}>
-                <Flamiliar />
+                <Flamiliar toggleFlamiliar={toggleFlamiliar}/>
               </Suspense>
             </div>
           </div>
@@ -663,187 +324,19 @@ function Chatroom() {
             >
               <div className="p-2">
                 <div className="flex justify-center aspect-w-16 aspect-h-9 relative aspect-video align-center ">
-                  {!isSuccess && (
-                    <div>
-                      <div>
-                        <VelocityScroll>LOADING GAME</VelocityScroll>
-                      </div>
-                      <div>
-                        <VelocityScroll>LOADING GAME</VelocityScroll>
-                      </div>
-                      <div className="flex justify-center">
-                        <img
-                          id="loading-image"
-                          src={loading}
-                          alt="Loading..."
-                        ></img>
-                      </div>
-                      <div>
-                        <VelocityScroll>LOADING GAME</VelocityScroll>
-                      </div>
-                    </div>
-                  )}
-                  {isSuccess && (
-                    <Application
-                      resizeTo={appRef}
-                      width={640}
-                      height={360}
-                      backgroundColor={' #FFFFFF'}
-                      resolution={2.5}
-                    >
-                      {mapPack.layers.map((objLay, layerIndex) => (
-                        <pixiContainer key={layerIndex}>
-                          {objLay.tiles.map((objTiles, index) => (
-                            <pixiSprite
-                              texture={Assets.get(
-                                nightClubTileSet[Math.floor(objTiles.id / 8)][
-                                  objTiles.id % 8
-                                ]
-                              )}
-                              ref={spriteRef2}
-                              x={32 * objTiles.x * 1.25}
-                              y={32 * objTiles.y * 1.25}
-                              scale={1.25}
-                              key={index}
-                            />
-                          ))}
-                        </pixiContainer>
-                      ))}
-                      {allPlayers.map((player) => {
-                        const getPlayerAnimation = () => {
-                          if (player.isSnapping) return snapTextures;
-                          if (player.isWalking) return walkTextures;
-                          if (player.isEnergyWaving) return energyWaveTextures;
-                          if (player.isWaving) return waveTextures;
-                          return textures; // Default standing animation
-                        };
-
-                        return (
-                          <pixiContainer
-                            x={player.x}
-                            y={player.y}
-                            key={player.id}
-                            scale={1.24}
-                          >
-                            {player.sentMessage && (
-                              <>
-                                <pixiSprite
-                                  texture={Assets.get('speech')}
-                                  key={player.username}
-                                  anchor={0.5}
-                                  x={70}
-                                  y={-30}
-                                  key={player.username}
-                                  scale={1.1}
-                                  width={player.currentMessage.length >= 70 ? 125: 110}
-                                  height={player.currentMessage.length >= 70 ? player.currentMessage.length >= 90 ? 140 : 110 : 70}
-                                />
-                                <pixiText
-                                  text={player.currentMessage}
-                                  anchor={0.5}
-                                  x={70}
-                                  y={-40}
-                                  key={player.username}
-                                  scale={1.1}
-                                  style={styleMessage}
-                                />
-                              </>
-                            )}
-
-                            <pixiText
-                              text={player.username}
-                              anchor={0.5}
-                              x={10}
-                              y={40}
-                              style={styleUserName}
-                            />
-                            <pixiSprite
-                              texture={Assets.get(player.username)}
-                              ref={spriteRef2}
-                              x={0}
-                              y={-13}
-                              scale={scaleFactor}
-                              width={25}
-                              height={25}
-                            ></pixiSprite>
-                            {isReady && (
-                              <pixiAnimatedSprite
-                                textures={getPlayerAnimation()}
-                                x={-18.6}
-                                y={-21}
-                                ref={(spriteRef) => spriteRef?.play()}
-                                initialFrame={0}
-                                animationSpeed={player.isSnapping ? 0.27 : 0.1}
-                                loop={true}
-                                scale={scaleFactor}
-                                width={64}
-                                height={64}
-                              />
-                            )}
-                            {isReady && player.isHearting && (
-                              <pixiAnimatedSprite
-                                textures={heartTextures}
-                                x={11}
-                                y={-48}
-                                rotation={0.5}
-                                ref={(spriteRef) => {
-                                  spriteRef?.play();
-                                }}
-                                initialFrame={0}
-                                animationSpeed={0.2}
-                                loop={true}
-                                scale={{
-                                  x: scaleFactor / 10,
-                                  y: scaleFactor / 10,
-                                }}
-                              />
-                            )}
-                            {/* Equipments & Emotes */}
-                            {isReady && (
-                              <>
-                                {player.equip420 && (
-                                  <pixiSprite
-                                    texture={Assets.get('joint')}
-                                    x={-1}
-                                    y={5}
-                                    width={10}
-                                    height={10}
-                                  />
-                                )}
-                                {player.equipBeer && (
-                                  <pixiSprite
-                                    texture={Assets.get('beer')}
-                                    x={-5}
-                                    y={5}
-                                    width={15}
-                                    height={15}
-                                  />
-                                )}
-                                {player.equipShades && (
-                                  <pixiSprite
-                                    texture={Assets.get('shades')}
-                                    x={2}
-                                    y={-9}
-                                    width={20}
-                                    height={20}
-                                  />
-                                )}
-                                {player.isSad && (
-                                  <pixiSprite
-                                    texture={Assets.get('sad')}
-                                    x={2}
-                                    y={-33}
-                                    width={20}
-                                    height={20}
-                                  />
-                                )}
-                              </>
-                            )}
-                          </pixiContainer>
-                        );
-                      })}
-                    </Application>
-                  )}
+                <Suspense fallback={ <div className="flex justify-center align-center">
+                            <img
+                              id="loading-image"
+                              src={loading}
+                              alt="Loading..."
+                            ></img>
+                            </div>}>
+                  <SocketContext.Provider value={socket}>
+                  <ChatroomContext.Provider value={eventId}>
+                   <MainChat onKeyboard={onKeyboard} chatSetOnKeyboard={chatSetOnKeyboard} avatarTextures={avatarTextures}/>
+                   </ChatroomContext.Provider>
+                   </SocketContext.Provider>
+                   </Suspense>
                 </div>
               </div>
             </div>
